+++ src/main/java/com/example/repositories/PeliculasRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entities.Peliculas;
 
-public interface PeliculasRepository extends JpaRepository<Peliculas,Long>{}
+public interface PeliculasRepository extends JpaRepository<Peliculas,Integer>{}
